+++ example/lib/main.dart
@@ -1,128 +1,108 @@
-// Copyright 2018 The Chromium Authors. All rights reserved.
-// Use of this source code is governed by a BSD-style license that can be
-// found in the LICENSE file.
-
-import 'dart:io';
-
-import 'package:device_info_plus/device_info_plus.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:location/location.dart';
-import 'package:maplibre_gl/maplibre_gl.dart';
-import 'package:maplibre_gl_example/attribution.dart';
-import 'package:maplibre_gl_example/get_map_informations.dart';
-import 'package:maplibre_gl_example/given_bounds.dart';
-import 'package:maplibre_gl_example/localized_map.dart';
+import 'package:flutter_web_plugins/url_strategy.dart';
+import 'package:maplibre_gl_example/annotation_circle_page.dart';
+import 'package:maplibre_gl_example/annotation_fill_page.dart';
+import 'package:maplibre_gl_example/annotation_layer_page.dart';
+import 'package:maplibre_gl_example/annotation_line_page.dart';
+import 'package:maplibre_gl_example/annotation_order_page.dart';
+import 'package:maplibre_gl_example/annotation_source_page.dart';
+import 'package:maplibre_gl_example/annotation_symbol_page.dart';
+import 'package:maplibre_gl_example/attribution_page.dart';
+import 'package:maplibre_gl_example/batch_operation_page.dart';
+import 'package:maplibre_gl_example/click_annotations_page.dart';
+import 'package:maplibre_gl_example/custom_marker_page.dart';
+import 'package:maplibre_gl_example/fullscreen_map_page.dart';
+import 'package:maplibre_gl_example/local_style_page.dart';
+import 'package:maplibre_gl_example/localized_map_page.dart';
+import 'package:maplibre_gl_example/main_page.dart';
+import 'package:maplibre_gl_example/map_state_page.dart';
+import 'package:maplibre_gl_example/move_camera_animated.dart';
+import 'package:maplibre_gl_example/move_camera_page.dart';
 import 'package:maplibre_gl_example/no_location_permission_page.dart';
+import 'package:maplibre_gl_example/offline_regions_page.dart';
+import 'package:maplibre_gl_example/scrolling_map_page.dart';
+import 'package:maplibre_gl_example/set_map_bounds_page.dart';
+import 'package:maplibre_gl_example/user_interface_page.dart';
+import 'package:maplibre_gl_example/various_sources_page.dart';
 
-import 'animate_camera.dart';
-import 'annotation_order_maps.dart';
-import 'click_annotations.dart';
-import 'custom_marker.dart';
-import 'full_map.dart';
-import 'layer.dart';
-import 'line.dart';
-import 'local_style.dart';
-import 'map_ui.dart';
-import 'move_camera.dart';
-import 'offline_regions.dart';
-import 'page.dart';
-import 'place_batch.dart';
-import 'place_circle.dart';
-import 'place_fill.dart';
-import 'place_source.dart';
-import 'place_symbol.dart';
-import 'scrolling_map.dart';
-import 'sources.dart';
+final routes = <String, WidgetBuilder>{
+  ExamplePage.main.path: (context) => const MainPage(),
+  ExamplePage.userInterface.path: (context) => const UserInterfacePage(),
+  ExamplePage.fullscreen.path: (context) => const FullscreenMapPage(),
+  ExamplePage.localized.path: (context) => const LocalizedMapPage(),
+  ExamplePage.moveCameraAnimated.path: (context) => const AnimateCameraPage(),
+  ExamplePage.moveCamera.path: (context) => const MoveCameraPage(),
+  ExamplePage.localStyle.path: (context) => const LocalStylePage(),
+  ExamplePage.scrolling.path: (context) => const ScrollingMapPage(),
+  ExamplePage.offlineRegions.path: (context) => const OfflineRegionsPage(),
+  ExamplePage.setMapBounds.path: (context) => const SetMapBoundsPage(),
+  ExamplePage.mapState.path: (context) => const MapStatePage(),
+  ExamplePage.noLocationPermission.path: (context) =>
+      const NoLocationPermissionPage(),
+  ExamplePage.annotationSymbol.path: (context) => const AnnotationSymbolPage(),
+  ExamplePage.annotationSource.path: (context) => const AnnotationSourcePage(),
+  ExamplePage.annotationLine.path: (context) => const AnnotationLinePage(),
+  ExamplePage.annotationLayer.path: (context) => const AnnotationLayerPage(),
+  ExamplePage.annotationCircle.path: (context) => const AnnotationCirclePage(),
+  ExamplePage.annotationFill.path: (context) => const AnnotationFillPage(),
+  ExamplePage.batchOperation.path: (context) => const BatchOperationPage(),
+  ExamplePage.annotationOrder.path: (context) => const AnnotationOrderPage(),
+  ExamplePage.customMarker.path: (context) => const CustomMarkerPage(),
+  ExamplePage.clickAnnotation.path: (context) => const ClickAnnotationPage(),
+  ExamplePage.variousSources.path: (context) => const VariousSourcesPage(),
+  ExamplePage.attribution.path: (context) => const AttributionPage(),
+};
 
-final List<ExamplePage> _allPages = <ExamplePage>[
-  const MapUiPage(),
-  const FullMapPage(),
-  const LocalizedMapPage(),
-  const AnimateCameraPage(),
-  const MoveCameraPage(),
-  const PlaceSymbolPage(),
-  const PlaceSourcePage(),
-  const LinePage(),
-  const LocalStylePage(),
-  const LayerPage(),
-  const PlaceCirclePage(),
-  const PlaceFillPage(),
-  const ScrollingMapPage(),
-  const OfflineRegionsPage(),
-  const AnnotationOrderPage(),
-  const CustomMarkerPage(),
-  const BatchAddPage(),
-  const ClickAnnotationPage(),
-  const Sources(),
-  const GivenBoundsPage(),
-  const GetMapInfoPage(),
-  const NoLocationPermissionPage(),
-  const AttributionPage(),
-];
+enum ExamplePage {
+  main('/', 'MapLibre Examples'),
+  userInterface('/user-interface', 'User interface'),
+  fullscreen('/fullscreen', 'Fullscreen map'),
+  localized('/localized', 'Localized map'),
+  moveCameraAnimated('/move-camera-animated', 'Move camera animated'),
+  moveCamera('/move-camera', 'Move camera'),
+  localStyle('/local-style', 'Local style'),
+  scrolling('/scrolling', 'Scrolling map'),
+  offlineRegions('/offline-regions', 'Offline regions'),
+  setMapBounds('/set-map-bounds', 'Set map bounds'),
+  mapState('/map-info', 'Get map state'),
+  noLocationPermission('/no-location-permission', 'No user location permission',
+      needsLocationPermission: false),
+  annotationSymbol('/annotation-symbol', 'Symbol'),
+  annotationSource('/annotation-source', 'Source'),
+  annotationLine('/annotation-line', 'Line'),
+  annotationLayer('/annotation-layer', 'Layer'),
+  annotationCircle('/annotation-circle', 'Circle'),
+  annotationFill('/annotation-fill', 'Fill'),
+  batchOperation('/batch-operation', 'Batch operation'),
+  annotationOrder('/annotation-order', 'Annotation order'),
+  customMarker('/custom-marker', 'Custom marker'),
+  clickAnnotation('/click-annotation', 'Click annotation'),
+  variousSources('/various-sources', 'Various sources'),
+  attribution('/attribution', 'Attribution');
 
-class MapsDemo extends StatefulWidget {
-  const MapsDemo({super.key});
+  const ExamplePage(
+    this.path,
+    this.title, {
+    this.needsLocationPermission = true,
+  });
 
-  @override
-  State<MapsDemo> createState() => _MapsDemoState();
-}
-
-class _MapsDemoState extends State<MapsDemo> {
-  /// Determine the android version of the phone and turn off HybridComposition
-  /// on older sdk versions to improve performance for these
-  ///
-  /// !!! Hybrid composition is currently broken do no use !!!
-  Future<void> initHybridComposition() async {
-    if (!kIsWeb && Platform.isAndroid) {
-      final androidInfo = await DeviceInfoPlugin().androidInfo;
-      final sdkVersion = androidInfo.version.sdkInt;
-      if (sdkVersion >= 29) {
-        MaplibreMap.useHybridComposition = true;
-      } else {
-        MaplibreMap.useHybridComposition = false;
-      }
-    }
-  }
-
-  void _pushPage(BuildContext context, ExamplePage page) async {
-    if (!kIsWeb && page.needsLocationPermission) {
-      final location = Location();
-      final hasPermissions = await location.hasPermission();
-      if (hasPermissions != PermissionStatus.granted) {
-        await location.requestPermission();
-      }
-    }
-    if (context.mounted) {
-      Navigator.of(context).push(MaterialPageRoute<void>(
-        builder: (_) => Scaffold(
-          appBar: AppBar(title: Text(page.title)),
-          body: page,
-        ),
-      ));
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(title: const Text('Maplibre examples')),
-      body: ListView.builder(
-        itemCount: _allPages.length + 1,
-        itemBuilder: (_, int index) => index == _allPages.length
-            ? const AboutListTile(
-                applicationName: "flutter-maplibre-gl example",
-              )
-            : ListTile(
-                leading: _allPages[index].leading,
-                title: Text(_allPages[index].title),
-                onTap: () => _pushPage(context, _allPages[index]),
-              ),
-      ),
-    );
-  }
+  final String path;
+  final String title;
+  final bool needsLocationPermission;
 }
 
 void main() {
-  runApp(const MaterialApp(home: MapsDemo()));
+  usePathUrlStrategy();
+  final materialTheme = ThemeData(
+    useMaterial3: true,
+    colorSchemeSeed: const Color(0xff295daa),
+  );
+
+  runApp(
+    MaterialApp(
+      theme: materialTheme,
+      initialRoute: ExamplePage.main.path,
+      routes: routes,
+    ),
+  );
 }