+++ example/lib/move_camera_animated.dart
@@ -4,27 +4,17 @@
 
 import 'package:flutter/material.dart';
 import 'package:maplibre_gl/maplibre_gl.dart';
+import 'package:maplibre_gl_example/common/example_scaffold.dart';
+import 'package:maplibre_gl_example/main.dart';
 
-import 'page.dart';
-
-class MoveCameraPage extends ExamplePage {
-  const MoveCameraPage({super.key})
-      : super(const Icon(Icons.map), 'Camera control');
-
-  @override
-  Widget build(BuildContext context) {
-    return const MoveCamera();
-  }
-}
-
-class MoveCamera extends StatefulWidget {
-  const MoveCamera({super.key});
+class AnimateCameraPage extends StatefulWidget {
+  const AnimateCameraPage({super.key});
 
   @override
-  State createState() => MoveCameraState();
+  State createState() => _AnimateCameraPageState();
 }
 
-class MoveCameraState extends State<MoveCamera> {
+class _AnimateCameraPageState extends State<AnimateCameraPage> {
   late MaplibreMapController mapController;
 
   void _onMapCreated(MaplibreMapController controller) {
@@ -33,55 +23,54 @@ class MoveCameraState extends State<MoveCamera> {
 
   @override
   Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: <Widget>[
-        Center(
-          child: SizedBox(
-            width: 300.0,
-            height: 200.0,
-            child: MaplibreMap(
-              onMapCreated: _onMapCreated,
-              onCameraIdle: () => debugPrint("onCameraIdle"),
-              initialCameraPosition:
-                  const CameraPosition(target: LatLng(0.0, 0.0)),
-            ),
-          ),
-        ),
-        Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-          children: <Widget>[
-            Column(
+    return ExampleScaffold(
+      page: ExamplePage.moveCameraAnimated,
+      body: Column(
+        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+        crossAxisAlignment: CrossAxisAlignment.stretch,
+        children: <Widget>[
+          Padding(
+            padding: const EdgeInsets.all(8),
+            child: Wrap(
+              alignment: WrapAlignment.center,
               children: <Widget>[
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
-                      CameraUpdate.newCameraPosition(
-                        const CameraPosition(
-                          bearing: 270.0,
-                          target: LatLng(51.5160895, -0.1294527),
-                          tilt: 30.0,
-                          zoom: 17.0,
-                        ),
-                      ),
-                    );
+                    mapController
+                        .animateCamera(
+                          CameraUpdate.newCameraPosition(
+                            const CameraPosition(
+                              bearing: 270.0,
+                              target: LatLng(51.5160895, -0.1294527),
+                              tilt: 30.0,
+                              zoom: 17.0,
+                            ),
+                          ),
+                        )
+                        .then(
+                          (result) => debugPrint(
+                              "mapController.animateCamera() returned $result"),
+                        );
                   },
                   child: const Text('newCameraPosition'),
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
-                      CameraUpdate.newLatLng(
-                        const LatLng(56.1725505, 10.1850512),
-                      ),
-                    );
+                    mapController
+                        .animateCamera(
+                          CameraUpdate.newLatLng(
+                            const LatLng(56.1725505, 10.1850512),
+                          ),
+                          duration: const Duration(seconds: 5),
+                        )
+                        .then((result) => debugPrint(
+                            "mapController.animateCamera() returned $result"));
                   },
                   child: const Text('newLatLng'),
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.newLatLngBounds(
                         LatLngBounds(
                           southwest: const LatLng(-38.483935, 113.248673),
@@ -97,7 +86,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.newLatLngZoom(
                         const LatLng(37.4231613, -122.087159),
                         11.0,
@@ -108,19 +97,15 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.scrollBy(150.0, -225.0),
                     );
                   },
                   child: const Text('scrollBy'),
                 ),
-              ],
-            ),
-            Column(
-              children: <Widget>[
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.zoomBy(
                         -0.5,
                         const Offset(30.0, 20.0),
@@ -131,7 +116,16 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
+                      CameraUpdate.newLatLngZoom(const LatLng(48, 11), 5),
+                      duration: const Duration(milliseconds: 300),
+                    );
+                  },
+                  child: const Text('latlngZoom'),
+                ),
+                TextButton(
+                  onPressed: () {
+                    mapController.animateCamera(
                       CameraUpdate.zoomBy(-0.5),
                     );
                   },
@@ -139,7 +133,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.zoomIn(),
                     );
                   },
@@ -147,7 +141,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.zoomOut(),
                     );
                   },
@@ -155,7 +149,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.zoomTo(16.0),
                     );
                   },
@@ -163,7 +157,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.bearingTo(45.0),
                     );
                   },
@@ -171,7 +165,7 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
                 TextButton(
                   onPressed: () {
-                    mapController.moveCamera(
+                    mapController.animateCamera(
                       CameraUpdate.tiltTo(30.0),
                     );
                   },
@@ -179,9 +173,17 @@ class MoveCameraState extends State<MoveCamera> {
                 ),
               ],
             ),
-          ],
-        )
-      ],
+          ),
+          Expanded(
+            child: MaplibreMap(
+              onMapCreated: _onMapCreated,
+              initialCameraPosition:
+                  const CameraPosition(target: LatLng(0.0, 0.0), zoom: 2),
+              styleString: "assets/osm_style.json",
+            ),
+          ),
+        ],
+      ),
     );
   }
 }